refactor: move resources::flavor types from lib to wire crate
