+++ lib/src/resources/flavor.rs
@@ -1,69 +1,14 @@
 use crate::common::{request, request_bare, SerializableNone};
 use crate::error::ApiError;
-use crate::resources::FlavorGroupMinimal;
 use anyhow::Context;
-use lrzcc_wire::common::display_option;
+use lrzcc_wire::resources::{
+    Flavor, FlavorCreateData, FlavorDetailed, FlavorImport, FlavorModifyData,
+    FlavorUsage, FlavorUsageAggregate,
+};
 use reqwest::blocking::Client;
 use reqwest::Url;
 use reqwest::{Method, StatusCode};
-use serde::{Deserialize, Serialize};
-use std::fmt::Display;
 use std::rc::Rc;
-use tabled::Tabled;
-
-#[derive(Clone, Debug, Deserialize, Serialize, Tabled)]
-pub struct Flavor {
-    pub id: u32,
-    pub name: String,
-    pub openstack_id: String, // UUIDv4
-    #[tabled(display_with = "display_option")]
-    pub group: Option<u32>,
-    #[tabled(display_with = "display_option")]
-    group_name: Option<String>,
-    pub weight: u32,
-}
-
-impl Display for Flavor {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str(&format!("Flavor(id={}, name={})", self.id, self.name))
-    }
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize, Tabled)]
-pub struct FlavorMinimal {
-    pub id: u32,
-    pub name: String,
-}
-
-// TODO maybe rethink the Display implementations
-impl Display for FlavorMinimal {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str(&format!("Flavor(id={}, name={})", self.id, self.name))
-    }
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize, Tabled)]
-pub struct FlavorDetailed {
-    pub id: u32,
-    pub name: String,
-    pub openstack_id: String, // UUIDv4
-    #[tabled(display_with = "display_option")]
-    pub group: Option<FlavorGroupMinimal>,
-    #[tabled(display_with = "display_option")]
-    pub group_name: Option<String>,
-    pub weight: u32,
-}
-
-impl Display for FlavorDetailed {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str(&format!("Flavor(id={}, name={})", self.id, self.name))
-    }
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize, Tabled)]
-pub struct FlavorImport {
-    pub new_flavor_count: u32,
-}
 
 pub struct FlavorApi {
     pub url: String,
@@ -124,27 +69,6 @@ impl FlavorListRequest {
     }
 }
 
-#[derive(Clone, Debug, Serialize)]
-struct FlavorCreateData {
-    name: String,
-    openstack_id: String, // UUIDv4
-    #[serde(skip_serializing_if = "Option::is_none")]
-    group: Option<u32>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    weight: Option<u32>,
-}
-
-impl FlavorCreateData {
-    fn new(name: String, openstack_id: String) -> Self {
-        Self {
-            name,
-            openstack_id,
-            group: None,
-            weight: None,
-        }
-    }
-}
-
 pub struct FlavorCreateRequest {
     url: String,
     client: Rc<Client>,
@@ -184,32 +108,6 @@ impl FlavorCreateRequest {
             Some(&self.data),
             StatusCode::CREATED,
         )
-    }
-}
-
-#[derive(Clone, Debug, Serialize)]
-struct FlavorModifyData {
-    id: u32,
-
-    #[serde(skip_serializing_if = "Option::is_none")]
-    name: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    openstack_id: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    group: Option<Option<u32>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    weight: Option<u32>,
-}
-
-impl FlavorModifyData {
-    fn new(id: u32) -> Self {
-        Self {
-            id,
-            name: None,
-            openstack_id: None,
-            group: None,
-            weight: None,
-        }
     }
 }
 
@@ -263,32 +161,6 @@ impl FlavorModifyRequest {
             StatusCode::OK,
         )
     }
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize, Tabled)]
-pub struct FlavorUsage {
-    pub user_id: u32,
-    pub user_name: String,
-    pub flavor_id: u32,
-    pub flavor_name: String,
-    #[tabled(display_with = "display_option")]
-    pub flavorgroup_id: Option<u32>,
-    #[tabled(display_with = "display_option")]
-    pub flavorgroup_name: Option<String>,
-    pub count: u32,
-    pub usage: u32,
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize, Tabled)]
-pub struct FlavorUsageAggregate {
-    pub flavor_id: u32,
-    pub flavor_name: String,
-    #[tabled(display_with = "display_option")]
-    pub flavorgroup_id: Option<u32>,
-    #[tabled(display_with = "display_option")]
-    pub flavorgroup_name: Option<String>,
-    pub count: u32,
-    pub usage: u32,
 }
 
 pub struct FlavorUsageRequest {
