+++ lib/src/resources/mod.rs
@@ -2,7 +2,6 @@ mod flavor;
 mod flavor_group;
 mod usage;
 
-// TODO rethink the public export of minimal structs
-pub use flavor::{FlavorApi, FlavorMinimal};
-pub use flavor_group::{FlavorGroupApi, FlavorGroupMinimal};
+pub use flavor::FlavorApi;
+pub use flavor_group::FlavorGroupApi;
 pub use usage::UsageApi;
