+++ wire/src/resources/mod.rs
@@ -1,0 +1,7 @@
+mod flavor;
+mod flavor_group;
+mod usage;
+
+pub use flavor::*;
+pub use flavor_group::*;
+pub use usage::*;
